requirements.txt
requests==2.31.0
pathvalidate==3.2.0
beautifulsoup4==4.12.3
urllib3==1.26.15
lxml==5.2.2
Jinja2==3.1.4
livereload==2.7.0
more-itertools==10.4.0
python-dotenv==1.0.0